requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-WTF==1.2.2
email-validator==2.2.0
passlib[argon2]==1.7.4
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.4
pytest-cov==6.0.0
